DDA-85: New public interface: compareAndMergeDiographs => move to own file + mock in generate-diograph-channel.spec.js
