+++ electron/channels/generate-diograph-channel.js
@@ -1,7 +1,7 @@
 const { generateDiograph } = require('../generators/diograph-generator')
 const { saveDiographJson } = require('../lib/save-diograph-json')
 const { readDiographJson } = require('../lib/read-diograph-json')
-const { isEmpty } = require('../lib/utils')
+const { compareAndMergeDiographs } = require('./compare-and-merge-diographs')
 
 /**
  * Event handler for GENERATE_DIOGRAPH channel
@@ -21,26 +21,12 @@ const { isEmpty } = require('../lib/utils')
  * }
  *
  */
-exports.generateDiographEventHandler = async (event, path) => {
+exports.generateDiographEventHandler = async function generateDiographEventHandler(event, path) {
   const existingDiograph = readDiographJson(path)
-  if (existingDiograph.diograph) {
-    const folderStructureDiograph = await generateDiograph(path)
-    const newDiories = compareDiographs(existingDiograph, folderStructureDiograph)
-    if (!isEmpty(newDiories)) {
-      const mergedDiograph = addDioriesToDiograph(newDiories, existingDiograph)
-      return mergedDiograph
-    }
-    return { ...existingDiograph, path }
-  }
-  const { rootId, diograph } = await generateDiograph(path)
-  await saveDiographJson(path, diograph, rootId)
-  return { rootId, diograph, path }
-}
-
-function compareDiographs(existingDiograph, folderStructureDiograph) {
-  return []
-}
-
-function addDioriesToDiograph(diories, diograph) {
-  return diories
+  const folderStructureDiograph = await generateDiograph(path)
+  const diograph = existingDiograph
+    ? compareAndMergeDiographs(existingDiograph, folderStructureDiograph)
+    : folderStructureDiograph
+  await saveDiographJson(path, diograph.diograph, diograph.rootId)
+  return { ...diograph, path }
 }
